Reformat and shit

+++ utils/ducks_config.py
@@ -154,4 +154,4 @@ def max_ducks_per_day(membercount):
     # 5000 members : https://www.wolframalpha.com/input/?i=plot+100+%2B+%28x+%2F+%285+%2B+x+%2F+450%29++%29+from+1+to+5000
     # 50000 members : https://www.wolframalpha.com/input/?i=plot+100+%2B+%28x+%2F+%285+%2B+x+%2F+450%29++%29+from+1+to+50000
 
-    return 100 + (membercount / (5 + membercount / 450))
+    return int(100 + (membercount / (5 + membercount / 450)))
